fix: provide a sensible example for a privateca Root CA example
This one looks a lot like someone copied by accident the subordinate
example out of `certificate_authority_subordinate/main.tf` as a root
CA. Thus it contains a lot of values set which are outright invalid
or not recommend for Root CA certficates if you consider RFC 5280
and CA/B Baseline Requirements as the standard to follow.

Also the subordinate example is a bit odd, e.g. configuring SAN
on any kind of CA certificate doesn't make sense. And the resources
examples there make use of the same pool name.

I tried to keep the lifetime setting, but set it to 99 years.
That is probably a sensible value for a P(rivate)KI setup. For
something public 10y or 15y are probably more sensible.

Signed-off-by: Sven Höxter <sven@stormbind.net>