Update main.tf
The sample that is currently featured in the GCF v2 docs does not work as is:

https://cloud.google.com/functions/docs/tutorials/terraform#create_your_maintf_file

The problem was flagged in this bug: b/343253268

The person who filed the bug updated the doc, but unfortunately they did not follow up with the sample fix. We got another bug about this recently: b/405057343

I have tested the fix and it now seems to work as expected.